some small performance improvements, and one potential big one. On line 48 (original) 54 in this version, the cache_get call is made multiple times (once for each object on the page perhaps) which is grinding the cache.  There is no need for it. It is attempting to address a race condition, but this should be solved not fixed the way it is.  This will remove multiple cache calls. Additionally, when the error condition is detected its removing the cache from the db using a direct db call which is probably incorrect as we are storing everything in memcache.
Around Line 14 we are checking if the dev tools module is there (it definitely is) so we can lose the if statement and we're also
using devtools_load() which is loading all of the tools, regardless if they are being used.  This should be refactored to only
load the tools we need.  I see that at line 27 we are loading the devtools_load('sbac_core') which then goes on to override the debugging
this should also be removed from production if it's not mission critical (i don't think it is).  Additionally, Appnovation external
ip address is hardcoded.  This is being loaded on each page request